Make the click to get out of the text editor work.
The deactivate function has a settimout to let the textEditor save the text before closing.
Otherwise, the deactivate function is called before any other function of the text editor, making it impossible to save the text.
Not calling the modifyText.setclicks(0) on deactivate make it bug when doing other operation on the text element like moving it.