Cxc 430 info of export format + CXC 191 Sharing functionality (#106)
* Change the text of the button

* Add the share function

The share function of Juju is now available. It can't be used on firefox since there is no browser compatibility.
An alert has been set to inform the user about it. It will be replaced by a toast as soons as the toast component is available.

* only the download button is pink

seen with alica, now only the download button is pink, the others are as secondary.

---------

Co-authored-by: Esitsolf <esitsolf@hotmail.fr>